refactor: refactor backend codebase
